Removed reporting subpackage, relocating classes and tests
