+++ cvpack/__init__.py
@@ -5,6 +5,7 @@
 """
 
 from ._version import __version__  # noqa: F401
+from .reporter import Reporter  # noqa: F401
 from .angle import Angle  # noqa: F401
 from .atomic_function import AtomicFunction  # noqa: F401
 from .attraction_strength import AttractionStrength  # noqa: F401
@@ -12,6 +13,7 @@ from .centroid_function import CentroidFunction  # noqa: F401
 from .collective_variable import CollectiveVariable  # noqa: F401
 from .composite_rmsd import CompositeRMSD  # noqa: F401
 from .distance import Distance  # noqa: F401
+from .global_parameter import GlobalParameter  # noqa: F401
 from .helix_angle_content import HelixAngleContent  # noqa: F401
 from .helix_hbond_content import HelixHBondContent  # noqa: F401
 from .helix_rmsd_content import HelixRMSDContent  # noqa: F401
