--- cvpack/reporting/__init__.py
@@ -1,6 +1,0 @@
-"""
-Reporting subpackage of CVPack
-
-"""
-
-from .reporting import CollectiveVariableReporter  # noqa: F401
